finalizada função reset de senha

+++ src/main/java/br/spin/crud/login/repository/LoginRepository.java
@@ -3,7 +3,7 @@ package br.spin.crud.login.repository;
 import br.spin.crud.login.models.Login;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginRepository extends JpaRepository<Login, Integer> {
+public interface LoginRepository extends JpaRepository<Login, Long> {
 
     public Login findByUsuario(String usuario);
